requirements.txt
streamlit
ultralytics
opencv-python-headless
pillow 
pandas
numpy
